disabled display of "empty" users, added "you don't have enough credits" msg for individual items.
items that are too expensive for the user no longer have a "buy" link.
users with no active items are no longer displayed. i suspect that i
too much logic into the haml files, might have to move that to the
controlles and leave only the display to the views.